refactor(st-09007): split react node test suites
